### 2023-07-23 #5684 @editorialbot
[PRE REVIEW]: TensorInference: A Julia package for tensor-based probabilistic inference
**Submitting author:** @mroavi (<a href="http://orcid.org/0009-0009-0291-503X">Martin Roa-Villescas</a>)
**Repository:** https://github.com/TensorBFS/TensorInference.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.2.0
