### 2023-07-23 1e4bacf
fix model printing
