+++ paper/paper.bib
@@ -79,7 +79,8 @@
   year={2022},
   eprint={2108.05665},
   archivePrefix={arXiv},
-  primaryClass={quant-ph}
+  primaryClass={quant-ph},
+  doi={10.48550/arXiv.2108.05665}
 }
 
 @article{gray2021hyper,
@@ -96,13 +97,19 @@
   pages={410}
 }
 
-@misc{pan2021simulating,
-  title={Simulating the Sycamore quantum supremacy circuits}, 
-  author={Feng Pan and Pan Zhang},
-  year={2021},
-  eprint={2103.03074},
-  archivePrefix={arXiv},
-  primaryClass={quant-ph}
+@article{Pan2022,
+  title = {Simulation of Quantum Circuits Using the Big-Batch Tensor Network Method},
+  author = {Pan, Feng and Zhang, Pan},
+  journal = {Phys. Rev. Lett.},
+  volume = {128},
+  issue = {3},
+  pages = {030501},
+  numpages = {6},
+  year = {2022},
+  month = {Jan},
+  publisher = {American Physical Society},
+  doi = {10.1103/PhysRevLett.128.030501},
+  url = {https://link.aps.org/doi/10.1103/PhysRevLett.128.030501}
 }
 
 @misc{liu2022computing,
@@ -125,7 +132,7 @@
   pages={273--288},
   year={2019},
   publisher={Oxford University Press},
-  url={https://doi.org/10.1093/imaiai/iay009}
+  doi={10.1093/imaiai/iay009}
 }
 
 @article{markov2008simulating,
@@ -137,7 +144,7 @@
   pages={963--981},
   year={2008},
   publisher={SIAM},
-  url={https://doi.org/10.1137/050644756}
+  doi={10.1137/050644756}
 }
 @article{orus2014practical,
 	doi = {10.1016/j.aop.2014.06.013},
@@ -149,7 +156,8 @@
 	pages = {117--158},
 	author = {Rom{\'{a}}n Or{\'{u}}s},
 	title = {A practical introduction to tensor networks: Matrix product states and projected entangled pair states},
-	journal = {Annals of Physics}
+	journal = {Annals of Physics},
+    doi={10.1016/j.aop.2014.06.013}
 }
 
 @article{orus2019tensor,
@@ -160,7 +168,8 @@
   number={9},
   pages={538--550},
   year={2019},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s42254-019-0086-7}
 }
 
 @inproceedings{roa2022partial,
@@ -192,7 +201,8 @@
   volume={28},
   number={2},
   pages={135--151},
-  year={2002}
+  year={2002},
+  doi={10.1145/567806.567807}
 }
 
 @article{cheng2019tree,
@@ -203,7 +213,8 @@
   number={15},
   pages={155131},
   year={2019},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevB.99.155131}
 }
 
 @article{han2018unsupervised,
@@ -214,14 +225,18 @@
   number={3},
   pages={031012},
   year={2018},
-  publisher={APS}
+  publisher={APS},
+  doi={10.1103/PhysRevX.8.031012}
 }
 
-@article{gao2021limitations,
-  title={Limitations of linear cross-entropy as a measure for quantum advantage},
-  author={Gao, Xun and Kalinowski, Marcin and Chou, Chi-Ning and Lukin, Mikhail D and Barak, Boaz and Choi, Soonwon},
-  journal={arXiv preprint arXiv:2112.01657},
-  year={2021}
+@misc{gao2021limitations,
[…]
