### 2023-07-23 #5684 @osorensen
:wave: @Jotho @Matt @Miles, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
