+++ dev/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"EditURL = \"https://github.com/TensorBFS/TensorInference.jl/blob/main/docs/src/generated/asia/main.jl\"","category":"page"},{"location":"generated/asia/main/#The-ASIA-network","page":"Asia network","title":"The ASIA network","text":"","category":"section"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"The graph below corresponds to the ASIA network, a simple Bayesian model used extensively in educational settings. It was introduced by Lauritzen in 1988 [lauritzen1988local].","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"┌───┐           ┌───┐\n│ A │         ┌─┤ S ├─┐\n└─┬─┘         │ └───┘ │\n  │           │       │\n  ▼           ▼       ▼\n┌───┐       ┌───┐   ┌───┐\n│ T │       │ L │   │ B │\n└─┬─┘       └─┬─┘   └─┬─┘\n  │   ┌───┐   │       │\n  └──►│ E │◄──┘       │\n      └─┬─┘           │\n┌───┐   │   ┌───┐     │\n│ X │◄──┴──►│ D │◄────┘\n└───┘       └───┘","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"The table below explains the meanings of each random variable used in the ASIA network model.","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"Random variable Meaning\nA Recent trip to Asia\nT Patient has tuberculosis\nS Patient is a smoker\nL Patient has lung cancer\nB Patient has bronchitis\nE Patient hast T and/or L\nX Chest X-Ray is positive\nD Patient has dyspnoea","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"We now demonstrate how to use the TensorInference.jl package for conducting a variety of inference tasks on the Asia network.","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"Import the TensorInference package, which provides the functionality needed for working with tensor networks and probabilistic graphical models.","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"using TensorInference","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"Load the ASIA network model from the asia.uai file located in the examples directory. See Model file format (.uai) for a description of the format of this file.","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"model = read_model_file(pkgdir(TensorInference, \"examples\", \"asia\", \"asia.uai\"))","category":"page"},{"location":"generated/asia/main/","page":"Asia network","title":"Asia network","text":"UAIModel(nvars = 8, nclique = 8)\n variables :\n factors : \n  Factor(1), size = (2,)\n  Factor(1, 2), size = (2, 2)\n  Factor(3), size = (2,)\n  Factor(3, 4), size = (2, 2)\n  Factor(3, 5), size = (2, 2)\n  Factor(2, 4, 6), size = (2, 2, 2)\n  Factor(6, 7), size = (2, 2)\n
[…]
