### 2023-07-27 #5684 @juliohm
Thank you for pinging. Unfortunately don't have the bandwidth to review
this time. Good luck on finding reviewers! :)

Em qui., 27 de jul. de 2023 15:27, Øystein Sørensen <
***@***.***> escreveu:

> 👋 @juliohm <https://github.com/juliohm>, @pitmonticone
> <https://github.com/pitmonticone>, @amadeusine
> <https://github.com/amadeusine>, would any of you be willing to review
> this submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
