### 2023-07-27 #5684 @emstoudenmire
I can help with this, but I am finding the context here a bit confusing, and what I'm asked to do. I see there is a link above about review criteria, but then it has a second link to an "example checklist" so is that link not a checklist? Is the paper I'm supposed to review in one of the links above? Which one? Where and how do I submit the review? Do I issue a command here? Thanks –
