### 2023-07-27 #5684 @osorensen
Thanks for responding, @emstoudenmire. This is a pre-review issue. Once I've found a sufficient number of reviewers, I'll start the actual review. There you will be able to generate your checklist and do the review. If interested, you can see an example of the review of a recently accepted paper here: https://github.com/openjournals/joss-reviews/issues/4561

I'll try to answer your questions here:
- You submit the review simply by filling out the checklist in the review issue. When there are issues that the authors need to fix, you are encouraged to open issues in the source repository, or by commenting in the review issue.
- The review concerns the whole submission, which consists of both a paper and accompanying software. I'll generate the pdf for you below this commment. You find the source repository for the software here: https://github.com/TensorBFS/TensorInference.jl
- You don't need to issue any commands here. When I start the review, there will be a short text describing the comment you need to issue.

Hope that was clarifying. In any case, you're welcome to reach out to me whenever you have further questions.
