### 2023-07-27 #5684 @gdalle
> I'm sorry @mroavi, I'm afraid I can't do that. That's something only editors are allowed to do.

Also I rather enjoyed how the bot put you in your place @mroavi :rofl:
