### 2023-07-27 #5700 @editorialbot
[REVIEW]: TensorInference: A Julia package for tensor-based probabilistic inference
**Submitting author:** @mroavi (<a href="http://orcid.org/0009-0009-0291-503X">Martin Roa-Villescas</a>)
**Repository:** https://github.com/TensorBFS/TensorInference.jl
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.4.1
