+++ src/TensorInference.jl
@@ -45,6 +45,7 @@ include("sampling.jl")
 using Requires
 function __init__()
     @require CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba" include("cuda.jl")
+    @require GenericTensorNetworks = "3521c873-ad32-4bb4-b63d-f4f178f42b49" include("generictensornetworks.jl")
 end
 
 # import PrecompileTools
