+++ Project.toml
@@ -19,6 +19,7 @@ TropicalNumbers = "b3a74e9c-7526-4576-a4eb-79c0d4c32334"
 [compat]
 CUDA = "4"
 DocStringExtensions = "0.8.6, 0.9"
+GenericTensorNetworks = "1"
 OMEinsum = "0.7"
 PrecompileTools = "1"
 Requires = "1"
