### 2023-08-15 53f2c36
new hard core lattice gas example
