### 2023-08-15 499316d
Example: hard core lattice gas (#58)
* new hard core lattice gas example

* suppress output

* polish examples
