### 2023-08-15 439af41
update_evidence! and update_temperature
