### 2023-08-17 3b6f51d
Return `Samples` type instead of a matrix (#60)
* sample type change

* update_evidence! and update_temperature
