+++ test/generictensornetworks.jl
@@ -12,6 +12,14 @@ using GenericTensorNetworks, TensorInference
     mars2 = TensorInference.normalize!(GenericTensorNetworks.solve(problem2, PartitionFunction(β)), 1)
     @test mars ≈ mars2
 
+    # update temperature
+    β2 = 3.0
+    model = update_temperature(model, problem, β2)
+    pa = probability(model)[]
+    model2 = TensorNetworkModel(problem, β2)
+    pb = probability(model2)[]
+    @test pa ≈ pb
+
     # mmap
     model = MMAPModel(problem, β; queryvars=[1,4])
     logp, config = most_probable_config(model)
