### 2023-09-03 #5700 @emstoudenmire
To the authors:

Overall this is an excellent submission. The need for the software is made clear in the writing, and the software is easy to install and appears to work as claimed. (I ran both of the example codes and the unit tests.)

Here are some points the authors should address before I approve of it:

* Would it be helpful to show typical example output of the code demo in the JOSS paper? Or is that not customary? I did note that the output is shown in the documentation.
* The documentation does not appear to give any Community Guidelines about how people can get involved in the project. I need to approve that these exists to finish the checklist above.
* One detail I noticed about the documentation is that the `marginals` function returns a set of 2-element arrays, but the docstring does not mention this or explain that aspect of the output. Is there a simple explanation? Could the docstring for `marginals` be expanded to explain this behavior?
