+++ paper/figures/asia-network/asia-network.tex
@@ -1,16 +1,72 @@
+% !TEX program = lualatex
+
 \documentclass[varwidth=14cm]{standalone}
 
 \usepackage{tikz}
 
+\usepackage[pdfa]{hyperref}
+\usepackage{hyperxmp}
+\hypersetup{%
+  unicode=true,
+  pdfapart=3,
+  pdfaconformance=B,
+  pdftitle={TensorInference: A Julia package for tensor-based
+  probabilistic inference},
+  pdfauthor={Martin Roa-Villescas, Jin-Guo Liu},
+  pdfpublication={Journal of Open Source Software},
+  pdfpublisher={Open Journals},
+  pdfissn={2475-9066},
+  pdfpubtype={journal},
+  pdfvolumenum={},
+  pdfissuenum={},
+  pdfdoi={10.xxxxxx/draft},
+  pdfcopyright={Copyright (c) 1970, Martin Roa-Villescas, Jin-Guo
+  Liu},
+  pdflicenseurl={http://creativecommons.org/licenses/by/4.0/},
+  colorlinks=true,
+  linkcolor=[rgb]{0.0, 0.5, 1.0},
+  citecolor=Blue,
+  urlcolor=[rgb]{0.0, 0.5, 1.0},
+  breaklinks=true
+}
+
 % Set fount used by JOSS papers
 \renewcommand\familydefault{\sfdefault}
+
+\ifnum 0\ifxetex 1\fi\ifluatex 1\fi=0 % if pdftex
+\usepackage[T1]{fontenc}
+\usepackage[utf8]{inputenc}
+
+\else % if luatex or xelatex
+  \ifxetex
+    \usepackage{mathspec}
+    \usepackage{fontspec}
+
+  \else
+    \usepackage{fontspec}
+  \fi
+  \defaultfontfeatures{Scale=MatchLowercase}
+  \defaultfontfeatures[\sffamily]{Ligatures=TeX}
+  \defaultfontfeatures[\rmfamily]{Ligatures=TeX,Scale=1}
+
+\fi
+% use upquote if available, for straight quotes in verbatim environments
+\IfFileExists{upquote.sty}{\usepackage{upquote}}{}
+% use microtype if available
+\IfFileExists{microtype.sty}{%
+  \usepackage{microtype}
+  \UseMicrotypeSet[protrusion]{basicmath} % disable protrusion for tt fonts
+}{}
+
+%% Font settings
+\usepackage{fontsetup} % Lazy way to get proper Greek lowercase glyphs
+
+% Use Hack https://sourcefoundry.org/hack/
+\setmonofont{Hack}
 
 % \fbox tip:
 % - https://tex.stackexchange.com/questions/119984/subfigures-side-by-side-with-captions
 \usepackage{subcaption}
-
-\usepackage{amsmath} % provides the \underset command
-\usepackage{amssymb} % provides the \sphericalangle symbol
 
 % tables generated by quarto-beamer use longtable
 \usepackage{booktabs}
