### 2023-09-04 38513c6
Move JunctionTrees' paragraph to 'Summary' section
