+++ paper/paper.md
@@ -157,101 +157,106 @@ recommend checking the
 directory of the official `TensorInference.jl` repository for the most
 up-to-date version of this example.
 
+```julia
+```
+
 ```{=latex}
 \DefineVerbatimEnvironment{Highlighting}{Verbatim}{commandchars=\\\{\},fontsize=\small}
 \newcommand{\InTok}[1]{\textcolor[rgb]{0.0, 0.0, 0.5}{#1}} % mrv
 \newcommand{\OutTok}[1]{\textcolor[rgb]{0.545, 0.0, 0.0}{#1}} % mrv
-```
 
-```julia
-# Import the TensorInference package, which provides the functionality needed
-# for working with tensor networks and probabilistic graphical models.
-using TensorInference
+\begin{Shaded}
+\begin{Highlighting}[]
+\CommentTok{\# Import the TensorInference package, which provides the functionality needed}
+\CommentTok{\# for working with tensor networks and probabilistic graphical models.}
+\InTok{In [1]: }\ImportTok{using} \BuiltInTok{TensorInference}
 
-# Load the ASIA network model from the `asia.uai` file located in the examples
-# directory. Refer to the documentation of this package for a description of the
-# format of this file.
-instance = read_instance(pkgdir(TensorInference, "examples", "asia", "asia.uai"))
+\CommentTok{\# Load the ASIA network model from the \textasciigrave{}asia.uai\textasciigrave{} file located in the examples}
+\CommentTok{\# directory. Refer to the documentation of this package for a description of the}
+\CommentTok{\# format of this file.}
+\NormalTok{instance }\OperatorTok{=} \FunctionTok{read\_instance}\NormalTok{(}\FunctionTok{pkgdir}\NormalTok{(TensorInference, }\StringTok{"examples"}\NormalTok{, }\StringTok{"asia"}\NormalTok{, }\StringTok{"asia.uai"}\NormalTok{))}
 
-# Create a tensor network representation of the loaded model.
-tn = TensorNetworkModel(instance)
+\CommentTok{\# Create a tensor network representation of the loaded model.}
+\NormalTok{tn }\OperatorTok{=} \FunctionTok{TensorNetworkModel}\NormalTok{(instance)}
 
-TensorNetworkModel{Int64, OMEinsum.DynamicNestedEinsum{Int64}, Array{Float64}}
-variables: 1, 2, 3, 4, 5, 6, 7, 8
-contraction time = 2^6.044, space = 2^2.0, read-write = 2^7.098
+\OutTok{Out [1]: }\NormalTok{TensorNetworkModel\{}\DataTypeTok{Int64}\NormalTok{, OMEinsum.DynamicNestedEinsum\{}\DataTypeTok{Int64}\NormalTok{\}, }\DataTypeTok{Array}\NormalTok{\{}\DataTypeTok{Float64}\NormalTok{\}\}}
+         \NormalTok{variables}\OperatorTok{:} \FloatTok{1}\NormalTok{, }\FloatTok{2}\NormalTok{, }\FloatTok{3}\NormalTok{, }\FloatTok{4}\NormalTok{, }\FloatTok{5}\NormalTok{, }\FloatTok{6}\NormalTok{, }\FloatTok{7}\NormalTok{, }\FloatTok{8}
[…]
