### 2023-09-04 972a4c8
docs: add a section on Why TensorInference.jl on the welcome page
