### 2023-09-04 18d8f92
docs: add the hard core lattice gas example to the examples overview
