+++ docs/src/examples-overview.md
@@ -3,6 +3,7 @@
 ```@contents
 Pages = [
   "generated/asia/main.md",
+  "generated/hard-core-lattice-gas/main.md",
 ]
 Depth = 1
 ```
