### 2023-09-04 709a022
docs: add a 'contributing' section
