### 2023-09-05 a066560
ci: use 'actions/checkout@v3' in favor of 'actions/checkout@v2'
