### 2023-09-05 365b689
let marginals return dict
