### 2023-09-05 6266d7c
Improve `marginals` docstring
