+++ paper/paper.md
@@ -172,13 +172,12 @@ up-to-date version of this example.
 \CommentTok{\# for working with tensor networks and probabilistic graphical models.}
 \InTok{In [1]: }\ImportTok{using} \BuiltInTok{TensorInference}
 
-\CommentTok{\# Load the ASIA network model from the \textasciigrave{}asia.uai\textasciigrave{} file located in the examples}
-\CommentTok{\# directory. Refer to the documentation of this package for a description of the}
-\CommentTok{\# format of this file.}
-\NormalTok{instance }\OperatorTok{=} \FunctionTok{read\_instance}\NormalTok{(}\FunctionTok{pkgdir}\NormalTok{(TensorInference, }\StringTok{"examples"}\NormalTok{, }\StringTok{"asia"}\NormalTok{, }\StringTok{"asia.uai"}\NormalTok{))}
+\CommentTok{\# Load the ASIA network model from \textasciigrave{}asia.uai\textasciigrave{} in the examples directory.}
+\CommentTok{\# Refer to the package documentation for a description of the format of this file.}
+\NormalTok{model }\OperatorTok{=} \FunctionTok{read\_model\_file}\NormalTok{(}\FunctionTok{pkgdir}\NormalTok{(TensorInference, }\StringTok{"examples"}\NormalTok{, }\StringTok{"asia"}\NormalTok{, }\StringTok{"asia.uai"}\NormalTok{))}
 
 \CommentTok{\# Create a tensor network representation of the loaded model.}
-\NormalTok{tn }\OperatorTok{=} \FunctionTok{TensorNetworkModel}\NormalTok{(instance)}
+\NormalTok{tn }\OperatorTok{=} \FunctionTok{TensorNetworkModel}\NormalTok{(model)}
 
 \OutTok{Out [1]: }\NormalTok{TensorNetworkModel\{}\DataTypeTok{Int64}\NormalTok{, OMEinsum.DynamicNestedEinsum\{}\DataTypeTok{Int64}\NormalTok{\}, }\DataTypeTok{Array}\NormalTok{\{}\DataTypeTok{Float64}\NormalTok{\}\}}
          \NormalTok{variables}\OperatorTok{:} \FloatTok{1}\NormalTok{, }\FloatTok{2}\NormalTok{, }\FloatTok{3}\NormalTok{, }\FloatTok{4}\NormalTok{, }\FloatTok{5}\NormalTok{, }\FloatTok{6}\NormalTok{, }\FloatTok{7}\NormalTok{, }\FloatTok{8}
@@ -192,70 +191,73 @@ up-to-date version of this example.
 \CommentTok{\# Calculate the marginal probabilities of each random variable in the model.}
 \InTok{In [3]: }\FunctionTok{marginals}\NormalTok{(tn)}
 
-\OutTok{Out [3]: }\FloatTok{8}\OperatorTok{{-}}\NormalTok{element }\DataTypeTok{Vector}\NormalTok{\{}\DataTypeTok{Vector}\NormalTok{\{}\DataTypeTok{Float64}\NormalTok{\}\}}\OperatorTok{:}
-        \NormalTok{ [}\FloatTok{0.01}\NormalTok{, }\FloatTok{0.99}\NormalTok{]}
[…]
