### 2023-09-05 #5700 @mroavi
Hi @emstoudenmire,

Thanks for your time and insightful feedback on our submission. We've incorporated your suggestions and added the missing parts. The paper is being regenerated to reflect these changes. You can check out the updated documentation on our [dev URL](https://tensorbfs.github.io/TensorInference.jl/dev/).
