### 2023-09-07 TensorBFS/TensorInference.jl#62 @gdalle
Asia network log partition is zero in the docs
Is that expected?

https://github.com/openjournals/joss-reviews/issues/5700
