### 2023-09-07 TensorBFS/TensorInference.jl#63 @gdalle
Dependencies on CUDA and GenericTensorNetworks could be package extensions
See https://pkgdocs.julialang.org/v1/creating-packages/#Conditional-loading-of-code-in-packages-(Extensions)

https://github.com/openjournals/joss-reviews/issues/5700
