### 2023-09-07 TensorBFS/TensorInference.jl#64 @gdalle
Precompilation is commented out
Is there a specific reason?

https://github.com/openjournals/joss-reviews/issues/5700
