### 2023-09-07 TensorBFS/TensorInference.jl#65 @gdalle
UAI reference comparison tests often fail
As a result they are commented out. Do we know why?

https://github.com/openjournals/joss-reviews/issues/5700
