### 2023-09-07 TensorBFS/TensorInference.jl#66 @gdalle
Tensor networks are never defined
Although I am familiar with the literature on graphical models, I didn't know what a tensor network was, or why the contraction order is important. I had to look up TensorOperations.jl, which in turn redirected me to [this paper](https://journals.aps.org/pre/abstract/10.1103/PhysRevE.90.033315) before I could understand this crucial aspect.

Both in the docs and in the JOSS paper, there needs to be an introduction to tensor networks and why they're harder to handle than it appears

https://github.com/openjournals/joss-reviews/issues/5700
