### 2023-09-07 TensorBFS/TensorInference.jl#67 @gdalle
State of the field is missing
The only explicit mention of alternatives provided in the JOSS paper is the [JunctionTrees.jl](https://github.com/mroavi/JunctionTrees.jl/) package, by the very same author. This is not enough to get an accurate picture of the field.

Some questions I have are:

- What are the leading packages for graphical model inference in Python, R and Julia?
- Which ones perform exact inference as opposed to variational inference, and why does it matter?
- How is TensorInference.jl different from the Julia competitors in terms of functionality and performance?
- Why would one choose the Julia option instead of Python or R?

A quick search yielded the following related Julia packages, but there are probably more:

- [RxInfer.jl](https://github.com/biaslab/RxInfer.jl)
- [GraphicalModelLearning.jl](https://github.com/lanl-ansi/GraphicalModelLearning.jl)
- [BayesNets.jl](https://github.com/sisl/BayesNets.jl)

Plus the heavyweight probabilistic programming languages like [Turing.jl](https://github.com/TuringLang/Turing.jl) and [Gen.jl](https://github.com/probcomp/Gen.jl)

https://github.com/openjournals/joss-reviews/issues/5700
