### 2023-09-07 TensorBFS/TensorInference.jl#68 @gdalle
Performance evaluation is missing
This package exists because tensor networks are assumed to be performant, but this claim is not backed by quantitative results.
I see that there are already benchmarks in the repo, would it be hard to compare them with other existing algorithms (at the very least JunctionTrees.jl)?

https://github.com/openjournals/joss-reviews/issues/5700
