### 2023-09-07 TensorBFS/TensorInference.jl#69 @gdalle
Quirks in the hard-core lattice gas example
- The $\infty$ sign mean in the energy formula might be a bit confusing, I would find it clearer to express the independent set constraint explicitly (unless the algorithm actually relies on this formulation with an objective barrier?)
- The lattice graphs are really cool but it's hard to distinguish shades of gray alone, could we maybe vary edge thickness too?
- The sampling step at the end returns an empty lattice

https://github.com/openjournals/joss-reviews/issues/5700
