### 2023-09-07 TensorBFS/TensorInference.jl#64 @GiggleLiu
Yes, the precompilation takes too long and breaks the CI tests.

We added precompilation in this PR:
https://github.com/TensorBFS/TensorInference.jl/pull/26

And removed it in the following PR:
https://github.com/TensorBFS/TensorInference.jl/pull/48

The reasons to remove precompilation are discussed in the above issues. We will add it back when the precompilation time is acceptable.
