+++ docs/make.jl
@@ -48,13 +48,16 @@ makedocs(;
     ),
     pages=[
         "Home" => "index.md",
-        "Background" => "background.md",
-        "Examples" => [
-            "Overview" => "examples-overview.md",
-            "Asia Network" => "generated/asia/main.md",
-            "Hard-core Lattice Gas" => "generated/hard-core-lattice-gas/main.md",
-           ],
-        "UAI file formats" => "uai-file-formats.md",
+        "Background" => [
+            "Probabilistic Inference" => "probabilisticinference.md",
+            "Tensor Networks" => "tensornetwork.md",
+            "UAI file formats" => "uai-file-formats.md"
+        ],
+        #"Examples" => [
+        #    "Overview" => "examples-overview.md",
+        #    "Asia Network" => "generated/asia/main.md",
+        #    "Hard-core Lattice Gas" => "generated/hard-core-lattice-gas/main.md",
+        #   ],
         "Performance tips" => "generated/performance.md",
         "API" => [
             "Public" => "api/public.md",
