+++ docs/src/tensornetwork.md
@@ -1,0 +1,211 @@
+# Tensor networks
+
+We now introduce the core ideas of tensor networks, highlighting their
+connections with the probabilistic graphical models (PGM) domain to align the terminology between them.
+
+For our purposes, a **tensor** is equivalent with the concept of a factor
+presented above, which we detail more formally below.
+
+## What is a tensor?
+*Definition*: A tensor $T$ is defined as:
+```math
+T: \prod_{V \in \bm{V}} \mathcal{D}_{V} \rightarrow \texttt{number}.
+```
+Here, the function $T$ maps each possible instantiation of the random
+variables in its scope $\bm{V}$ to a generic number type. In the context of tensor networks,
+a minimum requirement is that the number type is a commutative semiring.
+To define a commutative semiring with the addition operation $\oplus$ and the multiplication operation $\odot$ on a set $S$, the following relations must hold for any arbitrary three elements $a, b, c \in S$.
+```math
+\newcommand{\mymathbb}[1]{\mathbb{#1}}
+\begin{align*}
+(a \oplus b) \oplus c = a \oplus (b \oplus c) & \hspace{5em}\text{$\triangleright$ commutative monoid $\oplus$ with identity $\mymathbb{0}$}\\
+a \oplus \mymathbb{0} = \mymathbb{0} \oplus a = a &\\
+a \oplus b = b \oplus a &\\
+&\\
+(a \odot b) \odot c = a \odot (b \odot c)  &   \hspace{5em}\text{$\triangleright$ commutative monoid $\odot$ with identity $\mymathbb{1}$}\\
+a \odot  \mymathbb{1} =  \mymathbb{1} \odot a = a &\\
+a \odot b = b \odot a &\\
+&\\
+a \odot (b\oplus c) = a\odot b \oplus a\odot c  &  \hspace{5em}\text{$\triangleright$ left and right distributive}\\
+(a\oplus b) \odot c = a\odot c \oplus b\odot c &\\
+&\\
+a \odot \mymathbb{0} = \mymathbb{0} \odot a = \mymathbb{0}
+\end{align*}
+```
+Tensors are represented using multidimensional arrays of nonnegative numbers
+with labeled dimensions. These labels correspond to the array's indices, which
+in turn represent the set of random variables that the tensor is a function
+of. Thus, in this context, the terms **label**, **index**, and
+**variable** are synonymous and hence used interchangeably.
+
+## What is a tensor network?
+We now turn our attention to defining a **tensor network**.
+Tensor network a mathematical object that can be used to represent a multilinear map between tensors. It is widely used in condensed matter physics [^Orus2014][^Pfeifer2014] and quantum simulation [^Markov2008][^Pan2022]. It is also a powerful tool for solving combinatorial optimization problems [^Liu2023].
+It is important to note that we use a generalized version of the conventional
+notation, which is also knwon as the [eisnum](https://numpy.org/doc/stable/reference/generated/numpy.einsum.html) function that widely used in high performance computing.
+Packages that implement the conventional notation include
+- [numpy](https://numpy.org/doc/stable/reference/generated/numpy.einsum.html)
+- [OMEinsum.jl](https://github.com/under-Peter/OMEinsum.jl)
+- [PyTorch](https://pytorch.org/docs/stable/generated/torch.einsum.html)
+- [TensorFlow](https://www.tensorflow.org/api_docs/python/tf/einsum)
+
+This approach allows us to represent a more extensive set of sum-product multilinear operations between tensors, meeting the requirements of the PGM field.
+
[…]
