### 2023-09-10 TensorBFS/TensorInference.jl#68 @mroavi
Thanks for bringing this up, and you're absolutely right. Quantitative results are essential for backing up performance claims. We're already in the process of comparing our package against Merlin and libdai, which are two C++ libraries that have fared well in the UAI inference competition. These benchmarks should provide a good point of reference for the performance of our package.

Adding some of these results to the paper is a great idea. We'll make sure to include them to strengthen our claims about the package's performance. And yes, throwing in a comparison with JunctionTrees.jl could also be valuable. We'll look into it.
