### 2023-09-10 de14c1e
docs: make tensor diagrams larger
