+++ docs/src/tensornetwork.md
@@ -1,12 +1,14 @@
 # Tensor networks
 
 We now introduce the core ideas of tensor networks, highlighting their
-connections with the probabilistic graphical models (PGM) domain to align the terminology between them.
+connections with probabilistic graphical models (PGM) to align the terminology
+between them.
 
-For our purposes, a **tensor** is equivalent with the concept of a factor
-presented above, which we detail more formally below.
+For our purposes, a tensor is equivalent to the concept of a factor as defined
+in the PGM domain, which we detail more formally below.
 
 ## What is a tensor?
+
 *Definition*: A tensor $T$ is defined as:
 ```math
 T: \prod_{V \in \bm{V}} \mathcal{D}_{V} \rightarrow \texttt{number}.
@@ -39,17 +41,25 @@ of. Thus, in this context, the terms **label**, **index**, and
 **variable** are synonymous and hence used interchangeably.
 
 ## What is a tensor network?
-We now turn our attention to defining a **tensor network**.
-Tensor network a mathematical object that can be used to represent a multilinear map between tensors. It is widely used in condensed matter physics [^Orus2014][^Pfeifer2014] and quantum simulation [^Markov2008][^Pan2022]. It is also a powerful tool for solving combinatorial optimization problems [^Liu2023].
-It is important to note that we use a generalized version of the conventional
-notation, which is also knwon as the [eisnum](https://numpy.org/doc/stable/reference/generated/numpy.einsum.html) function that widely used in high performance computing.
-Packages that implement the conventional notation include
+
+We now turn our attention to defining a **tensor network**, a mathematical
+object used to represent a multilinear map between tensors. This concept is
+widely employed in fields like condensed matter physics
+[^Orus2014][^Pfeifer2014], quantum simulation [^Markov2008][^Pan2022], and
+even in solving combinatorial optimization problems [^Liu2023]. It's worth
+noting that we use a generalized version of the conventional notation, most
+commonly known through the
+[eisnum](https://numpy.org/doc/stable/reference/generated/numpy.einsum.html)
+function, which is commonly used in high-performance computing. Packages that
+implement this conventional notation include
 - [numpy](https://numpy.org/doc/stable/reference/generated/numpy.einsum.html)
 - [OMEinsum.jl](https://github.com/under-Peter/OMEinsum.jl)
 - [PyTorch](https://pytorch.org/docs/stable/generated/torch.einsum.html)
 - [TensorFlow](https://www.tensorflow.org/api_docs/python/tf/einsum)
 
-This approach allows us to represent a more extensive set of sum-product multilinear operations between tensors, meeting the requirements of the PGM field.
+This approach allows us to represent a broader range of sum-product
+multilinear operations between tensors, thus meeting the requirements of the
+PGM field.
 
 *Definition*[^Liu2023]: A tensor network is a multilinear map represented by the triple
 $\mathcal{N} = (\Lambda, \mathcal{T}, \bm{\sigma}_0)$, where:
@@ -62,8 +72,8 @@ $\mathcal{N} = (\Lambda, \mathcal{T}, \bm{\sigma}_0)$, where:
 
 More specifically, each tensor $T^{(k)}_{\bm{\sigma}_k} \in \mathcal{T}$ is
 labeled by a string $\bm{\sigma}_k \in \Lambda^{r \left(T^{(k)} \right)}$, where
-$r \left(T^{(k)} \right)$ is the rank of $T^{(k)}$. The multilinear map, or
-the `contraction`, applied to this triple is defined as
+$r \left(T^{(k)} \right)$ is the rank of $T^{(k)}$. The multilinear map, also
+known as the `contraction`, applied to this triple is defined as
[…]
