### 2023-09-10 TensorBFS/TensorInference.jl#62 @mroavi
The partition function \( Z \) is basically the 'normalizer' for our model. It ensures that the probabilities sum to 1. If you've already got all your factors normalized, like in our toy example, then \( Z \) will be just 1.

We've added a comment in the example to clarify this point.
