### 2023-09-10 TensorBFS/TensorInference.jl#69 @mroavi
Thanks for the suggestions! We've updated the energy formula, fixed the empty lattice issue, and tweaked the lattice graph's edge thickness. It now looks like this:

![image](https://github.com/TensorBFS/TensorInference.jl/assets/19517248/aabbccd3-4fff-4dc6-89df-fc0151b5c537)
