### 2023-09-10 TensorBFS/TensorInference.jl#66 @mroavi
Absolutely, glad you pointed that out. We've added a whole new section in the documentation to introduce tensor networks and explain their connection with probabilistic graphical models. You can check it out here:
https://tensorbfs.github.io/TensorInference.jl/dev/tensornetwork/.
