### 2023-09-11 TensorBFS/TensorInference.jl#71 @gdalle
How to free some room in the paper?
Opening this issue to keep track of what to cut in order to add details for #66 #67 #68

https://github.com/openjournals/joss-reviews/issues/5700
