### 2023-09-11 0147ad6
Minor improvements to the ASIA network example
