### 2023-09-11 b8f6c3a
Remove numbers from the equations in the hard-core lattice gas example
These don't display properly.
