### 2023-09-11 d715d66
Addresses https://github.com/TensorBFS/TensorInference.jl/issues/62
