+++ paper/paper.md
@@ -1,10 +1,6 @@
 ---
 title: 'TensorInference: A Julia package for tensor-based probabilistic
 inference'
-#title: 'TensorInference: A Julia package for probabilistic inference through
-#tensor-based technology'
-#title: 'TensorInference: A Julia package for probabilistic inference using
-#tensor networks'
 tags:
   - Julia
   - probabilistic graphical models
