### 2023-09-11 TensorBFS/TensorInference.jl#66 @gdalle
Cool addition! I'll give some thought to what we can trim from the paper
