+++ paper/paper.bib
@@ -306,3 +306,38 @@
   doi          = {10.5281/zenodo.8166121},
   url          = {https://doi.org/10.5281/zenodo.8166121}
 }
+
+@online{marinescu2022merlin,
+  author = {Radu Marinescu},
+  title = {Merlin},
+  year = 2022,
+  url = {\url{https://www.ibm.com/opensource/open/projects/merlin/} [Accessed: 11 September 2023]},
+  urldate = {2023-09-11}
+}
+
+@article{mooij2010libdai,
+  author    = {Joris M. Mooij},
+  title     = {lib{DAI}: A Free and Open Source {C++} Library for Discrete Approximate Inference in Graphical Models},
+  journal   = {Journal of Machine Learning Research},
+  year      = 2010,
+  month     = Aug,
+  volume    = 11,
+  pages     = {2169-2173},
+  url       = "http://www.jmlr.org/papers/volume11/mooij10a/mooij10a.pdf"
+}
+
+@online{gal2010summary,
+  title = {Summary of the 2010 {UAI} approximate inference challenge},
+  year = 2010,
+  howpublished = {\url{https://www.cs.huji.ac.il/project/UAI10/summary.php} [Accessed: 11 September 2023]},
+  urldate = {2021-08-21},
+  author = {Elidan, Gal and Globerson, Amir},
+}
+
+@online{gogate2014uai,
+  title = {{UAI} 2014 {Probabilistic} {Inference} {Competition}},
+  year = 2014,
+  howpublished = {\url{https://www.ics.uci.edu/~dechter/softwares/benchmarks/Uai14/UAI_2014_Inference_Competition.pdf} [Accessed: 11 September 2023]},
+  urldate = {2023-09-11},
+  author = {Gogate, Vibhav},
+}
