+++ paper/scripts/performance-evaluation/generate-graph.jl
@@ -1,0 +1,213 @@
+# Color palettes: 
+# - https://github.com/matplotlib/matplotlib/issues/9460#issuecomment-875185352
+# - https://personal.sron.nl/~pault/#tab:blindvision
+# - https://yoshke.org/blog/colorblind-friendly-diagrams
+# - https://davidmathlogic.com/colorblind
+# - https://www.color-hex.com/color-palette/1018347
+
+using PGFPlotsX, CSV, DataFrames, Artifacts, StatsBase
+using JunctionTrees: get_td_soln
+
+push!(
+  PGFPlotsX.CUSTOM_PREAMBLE,
+  """
+\\usepackage[T1]{fontenc}
+\\usepackage{xcolor}
+\\usepackage{tikz}
+\\usepackage{pgfplots}
+\\usepackage{amsmath,amssymb}
+% Bright qualitative colour scheme that is colour-blind safe
+% https://personal.sron.nl/~pault/#tab:blindvision
+\\definecolor{c01}{HTML}{4477AA}
+\\definecolor{c02}{HTML}{EE6677}
+\\definecolor{c03}{HTML}{228833}
+\\definecolor{c04}{HTML}{CCBB44}
+\\definecolor{c05}{HTML}{66CCEE}
+\\definecolor{c06}{HTML}{AA3377}
+\\definecolor{c07}{HTML}{BBBBBB}
+\\definecolor{c08}{HTML}{BBBBBB}
+\\usepackage{fontsetup}
+\\setmonofont{Hack}
+"""
+)
+
+# data_file = ARGS[1]
+
+# DEBUG
+data_file = "./out/co23/2023-09-10--20-20-45/out.csv"
+
+df1 =
+  CSV.File(data_file) |> # read benchmark data from file
+  DataFrame |> # convert it to a data_file frame
+   x -> unstack(x, :problem, :library, :execution_time) |> # convert from long to wide format (create a column for each possible `library` value)
+   dropmissing # drop rows with missing values
+
+df2 =
+  map(x -> joinpath(artifact"uai2014-mar", x * ".tamaki.td"), df1.problem) |> # create absolute filepaths for each problem in `df`
+  x -> get_td_soln.(x) |> # get the tree decomposition solution line ([nbags, :largest_bag_size, nvars]) for each problem 
+  x -> DataFrame(x=[i for i in x]) |> # create data frame using the constructor for vector of vectors
+  x -> select(x, :x => AsTable) |> # https://www.juliabloggers.com/handling-vectors-of-vectors-in-dataframes-jl/
+  x -> rename(x, [:nbags, :largest_bag_size, :nvars]) # rename columns
+
+df =
+  hcat(df1, df2) |> # horizontally concat data frame
+  x -> sort(x, [:largest_bag_size, :nvars, :nbags]) |> # sort the rows based on the largest bag size
+  x -> transform(x, [:libdai, :TensorInference] => (./) => :libdai_ti_speedup) |>
+  x -> transform(x, [:merlin, :TensorInference] => (./) => :merlin_ti_speedup) |>
+  x -> transform(x, [:JunctionTrees_v1, :TensorInference] => (./) => :jtv1_ti_speedup) |>
+  x -> transform(x, [:JunctionTrees_v2, :TensorInference] => (./) => :jtv2_ti_speedup)
+
+labels =
+  df.problem |>
+  x -> match.(r"[a-zA-Z]+", x) |>
+  x -> getfield.(x, :match)
+
+labels_unique = unique(labels) |> sort
+xmax = maximum(df.largest_bag_size) + 1
+
+@pgf tp = Axis(
+  {
+    # title="TensorInference.jl Speedup",
+    xmin = 0,
[…]

[…]
