+++ docs/src/examples-overview.md
@@ -2,7 +2,7 @@
 
 ```@contents
 Pages = [
-  "generated/asia/main.md",
+  "generated/asia-network/main.md",
   "generated/hard-core-lattice-gas/main.md",
 ]
 Depth = 1
