### 2023-09-11 eb27f4c
fix tensor network doc
