### 2023-09-11 dce0ec7
docs: add a section on performance evaluation
