### 2023-09-12 c250935
docs: generate performance evaluation graph with Julia
and move benchmark code to https://github.com/TensorBFS/TensorInferenceBenchmarks.jl
