### 2023-09-12 bdad795
Improve performance evaluation graph
