+++ docs/src/performance-evaluation.md
@@ -3,998 +3,19 @@
 The graph below illustrates a comparison of the runtime performance of
 *TensorInference.jl* against *Merlin* [^marinescu2022merlin], *libDAI*
 [^mooij2010libdai], and *JunctionTrees.jl* [^roa2022partial][^roa2023scaling]
-libraries. Both *Merlin* and *libDAI* have previously participated in UAI
-inference competitions [^gal2010summary][^gogate2014uai], achieving favorable
-results. Additionally, we compared against two versions of *JunctionTrees.jl*,
-the predecessor of *TensorInference.jl*. The first version does not employ
-tensor technology, while the second version optimizes individual sum-product
-computations using tensor-based technology. The experiments were conducted on
-an Intel Core i9--9900K CPU @3.60GHz with 64 GB of RAM.
-
-```@eval
-using TikzPictures
-
-tp = TikzPicture(
-  L"""
-  \begin{axis}[
-    axis line style={black!100},
-    every axis label/.append style={black!100},
-    every tick label/.append style={black!100},
-    xmin={0},
-    xmax={23},
-    xlabel={Largest cluster size},
-    xmajorgrids={true},
-    ymin={0},
-    ymax={1000000},
-    ymode={log},
-    ytick={0.001,0.01,0.1,1.0,10.0,100.0,1000.0,10000.0,100000.0},
-    ymajorgrids={true},
-    ylabel={Run time speedup},
-    label style={font={\footnotesize}},
-    tick label style={font={\footnotesize}},
-    scatter/classes={Alchemy={mark={x}}, CSP={mark={+}}, Grids={mark={asterisk}}, ObjectDetection={mark={-}}, Pedigree={mark={triangle}}, Promedus={mark={o}}, Segmentation={mark={Mercedes star}}, linkage={mark={diamond}}},
-    legend style={legend columns={3}, at={(0.51,-0.4)}, anchor={south}, text={black!100}, draw={none}, fill={none}, font={\footnotesize}, column sep={1.5}}
-    ]
-    \addplot[
-      c01,
-      scatter,
-      only marks,
-      scatter src={explicit symbolic},
-      legend image post style={black!100},
-      legend style={text={black!100}, font={\footnotesize}}
-    ]
-      table[row sep={\\}, meta={label}]
-      {
-        x  y  label  \\
-        4  2.8892732905905696  Promedus
-        \\
-        5  0.9992454677080572  Promedus
-        \\
-        5  2.7291490240125063  Promedus
-        \\
-        6  162.92710213915234  ObjectDetection
-        \\
-        6  143.18430666866695  ObjectDetection
-        \\
-        6  720.0876395098279  ObjectDetection
-        \\
-        6  217.1158405786517  ObjectDetection
-        \\
-        6  147.2467303223847  ObjectDetection
-        \\
-        6  170.3349686317474  ObjectDetection
-        \\
-        6  734.1273502948045  ObjectDetection
-        \\
-        6  144.59866923029833  ObjectDetection
-        \\
-        6  728.2931567239458  ObjectDetection
-        \\
-        6  138.79291257761554  ObjectDetection
-        \\
-        6  122.47593563280027  ObjectDetection
-        \\
-        6  139.85747966450083  ObjectDetection
-        \\
-        6  131.525522176854  ObjectDetection
-        \\
-        6  109.92499089134364  ObjectDetection
-        \\
-        6  93.72644489626983  ObjectDetection
-        \\
-        6  98.16016046784863  ObjectDetection
-        \\
-        6  119.60310364630723  ObjectDetection
-        \\
-        6  99.38321389928038  ObjectDetection
-        \\
-        6  111.86522014674959  ObjectDetection
-        \\
[…]
