+++ docs/src/performance-evaluation.md
@@ -2,17 +2,14 @@
 
 The graph below illustrates a comparison of the runtime performance of
 *TensorInference.jl* against *Merlin* [^marinescu2022merlin], *libDAI*
-[^mooij2010libdai], and *JunctionTrees.jl* [^roa2022partial][^roa2023scaling]
-libraries, specifically for the task of computing the marginal probabilities
-of all variables. Both *Merlin* and *libDAI* have previously participated in
-UAI inference competitions [^gal2010summary][^gogate2014uai], achieving
-favorable results. Additionally, we compared against two versions of
-*JunctionTrees.jl*, the predecessor of *TensorInference.jl*. The first version
-does not employ tensor technology, while the second version optimizes
-individual sum-product computations using tensor-based technology. The
-experiments were conducted on an Intel Core i9--9900K CPU @3.60GHz with 64 GB
-of RAM. Performance comparisons for other tasks will be added in the near
-future.
+[^mooij2010libdai], and *JunctionTrees.jl* [^roa2022partial] libraries,
+specifically for the task of computing the marginal probabilities of all
+variables. Both *Merlin* and *libDAI* have previously participated in UAI
+inference competitions [^gal2010summary][^gogate2014uai], achieving favorable
+results. Additionally, we compared against *JunctionTrees.jl*, the predecessor
+of *TensorInference.jl*. The experiments were conducted on an Intel Core
+i9--9900K CPU @3.60GHz with 64 GB of RAM. Performance comparisons for other
+tasks will be added in the near future.
 
 ```@raw html
 <img src="./assets/performance-evaluation.svg" width="700" style="margin-left: auto; margin-right: auto; display:block;">
@@ -20,9 +17,15 @@ future.
 
 The benchmark problems are arranged along the x-axis in ascending order of
 complexity, measured by the induced tree width. On average,
-*TensorInference.jl* achieves a speedup of 11 times across all problems.
-Notably, for the 10 most complex problems, the average speedup increases to 63
-times, highlighting its superior scalability.
+*TensorInference.jl* achieves a speedup of 20 times across all problems.
+Notably, for the 10 most complex problems, the average speedup increases to
+148 times, highlighting its superior scalability. The graph features a fitted
+linear curve in log-space to underscore the exponential improvement in
+computation time achieved by *TensorInference.jl* in comparison to the other
+alternatives. This speedup is primarily due to our package's unique approach:
+while traditional solvers typically focus only on minimizing space complexity
+(as quantified by the induced tree width), *TensorInference.jl* is
+designed to optimize for both time and space complexity. 
 
 ## References
 
@@ -45,11 +48,6 @@ times, highlighting its superior scalability.
         Machine Learning Research*, 11:2169-2173, Aug 2010. [Online].
         Available:
         [http://www.jmlr.org/papers/volume11/mooij10a/mooij10a.pdf](http://www.jmlr.org/papers/volume11/mooij10a/mooij10a.pdf).
-
-[^roa2023scaling]:
-    Martin Roa-Villescas, Jin Guo Liu, Patrick W.A. Wijnings, Sander Stuijk, Henk Corporaal.
-    "Scaling Probabilistic Inference through Message Contraction Optimization". 
-    In: 2023 Congress in Computer Science, Computer Engineering, & Applied Computing (CSCE), 2023.
 
 [^roa2022partial]:
     Martin Roa-Villescas, Patrick W.A. Wijnings, Sander Stuijk, Henk Corporaal. 
