### 2023-09-12 bdae7b2
docs: add a section on performance evaluation (#74)
* docs: add a section on performance evaluation

* docs: generate performance evaluation graph with Julia

and move benchmark code to https://github.com/TensorBFS/TensorInferenceBenchmarks.jl

* Improve performance evaluation graph

* docs: update the section and graph on performance evaluation

* docs: improve ambiguous file names
