### 2023-09-12 15e076c
docs: fix references in performance evaluation section
