### 2023-09-12 TensorBFS/TensorInference.jl#68 @mroavi
Hi @gdalle . We just wrapped up a performance comparison against three other libraries: *Merlin*, *libDAI*, and, of course, *JunctionTrees.jl*. Merlin and libDAI are C++ libraries that have historically performed well in UAI inference competitions. JunctionTrees.jl is essentially the predecessor of TensorInference.jl and is also implemented in Julia. You can check out all the benchmark results on our docs dev site: [TensorInference Performance Evaluation](https://tensorbfs.github.io/TensorInference.jl/dev/performance-evaluation/).
