### 2023-09-12 83ee4e7
docs: fix unresolved relative path to performance evaluation image
