### 2023-09-14 TensorBFS/TensorInference.jl#71 @gdalle
Here are a few sentences I think are dispensable:

> The advantage of supporting generic element types lies in the
ability to solve a variety of problems using the same tensor network
contraction algorithm, simply by varying the element types used.

> A major challenge in developing intelligent systems is the ability to reason
under uncertainty, a challenge that appears in many real-world problems across
various domains, including artificial intelligence, medical diagnosis, computer
vision, computational biology, and natural language processing.

> This process, which can be
thought of as drawing global insights from local observations, is known as
*probabilistic inference*.

> Additionally, they
form the foundation for various algorithms that enable efficient probabilistic
inference.

> To tackle these
challenges, more efficient and scalable inference algorithms are needed.

I would also shorten the usage example
