### 2023-09-16 00f065a
Address the issues raised by gdalle regarding ...
Exact vs. Approximate inference
Performance evaluation
Alternative packages for probabilistic inference
