+++ paper/paper.bib
@@ -306,3 +306,86 @@
   doi          = {10.5281/zenodo.8166121},
   url          = {https://doi.org/10.5281/zenodo.8166121}
 }
+
+@article{carpenter2017stan,
+  title={Stan: A probabilistic programming language},
+  author={Carpenter, Bob and Gelman, Andrew and Hoffman, Matthew D and Lee, Daniel and Goodrich, Ben and Betancourt, Michael and Brubaker, Marcus and Guo, Jiqiang and Li, Peter and Riddell, Allen},
+  journal={Journal of statistical software},
+  volume={76},
+  number={1},
+  year={2017},
+  publisher={Columbia Univ., New York, NY (United States); Harvard Univ., Cambridge, MA (United States)}
+}
+
+@article{oriol2023pymc,
+  title={PyMC: A Modern and Comprehensive Probabilistic Programming Framework in Python},
+  author={Abril-Pla Oriol and Andreani Virgile and Carroll Colin and Dong Larry and Fonnesbeck Christopher J. and Kochurov Maxim and Kumar Ravin and Lao Jupeng and Luhmann Christian C. and Martin Osvaldo A. and Osthege Michael and Vieira Ricardo and Wiecki Thomas and Zinkov Robert},
+  journal = {{PeerJ} Computer Science},
+  publisher = {{PeerJ}},
+  volume={9},
+  pages={e1516},
+  year={2023},
+  doi={10.7717/peerj-cs.1516}
+}
+
+@inproceedings{ge2018turing,
+  author    = {Hong Ge and
+               Kai Xu and
+               Zoubin Ghahramani},
+  title     = {Turing: a language for flexible probabilistic inference},
+  booktitle = {International Conference on Artificial Intelligence and Statistics,
+               {AISTATS} 2018, 9-11 April 2018, Playa Blanca, Lanzarote, Canary Islands,
+               Spain},
+  pages     = {1682--1690},
+  year      = {2018},
+  url       = {http://proceedings.mlr.press/v84/ge18b.html},
+  biburl    = {https://dblp.org/rec/bib/conf/aistats/GeXG18},
+}
+
+@article{bagaev2023rxinfer,
+  doi = {10.21105/joss.05161},
+  url = {https://doi.org/10.21105/joss.05161},
+  year = {2023},
+  publisher = {The Open Journal},
+  volume = {8},
+  number = {84},
+  pages = {5161},
+  author = {Dmitry Bagaev and Albert Podusenko and Bert de Vries},
+  title = {RxInfer: A Julia package for reactive real-time Bayesian inference},
+  journal = {Journal of Open Source Software}
+}
+
+@article{mooij2010libdai,
+  author    = {Joris M. Mooij},
+  title     = {lib{DAI}: A Free and Open Source {C++} Library for Discrete Approximate Inference in Graphical Models},
+  journal   = {Journal of Machine Learning Research},
+  year      = 2010,
+  month     = Aug,
+  volume    = 11,
+  pages     = {2169-2173},
+  url       = "http://www.jmlr.org/papers/volume11/mooij10a/mooij10a.pdf"
+}
+
+@misc{marinescu2022merlin,
+  author = {Radu Marinescu},
+  title = {Merlin},
+  year = 2022,
+  note = {\url{https://www.ibm.com/opensource/open/projects/merlin/}},
+  urldate = {2022-02-25}
+}
+
+@article{hurley2016multi,
+  TITLE = {{Multi-language evaluation of exact solvers in graphical model discrete optimization}},
[…]
