### 2023-09-16 TensorBFS/TensorInference.jl#66 @mroavi
The paper now includes a short introduction to tensors, contractions, and tensor networks, and it explains why the contraction order is important.
