### 2023-09-16 TensorBFS/TensorInference.jl#68 @mroavi
I added a reference in the paper to the performance evaluation section in the docs.
