### 2023-09-16 TensorBFS/TensorInference.jl#67 @mroavi
We included a discussion of the trade-off between exact and approximate inference in the paper and explained why we opted for an exact approach. The paper now mentions renowned packages for both approximate and exact inference. We also added a reference to the performance evaluation in our package's documentation, which compares TensorInference.jl with its predecessor JunctionTrees.jl, as well as with Merlin and libDAI.
