### 2023-09-16 TensorBFS/TensorInference.jl#71 @mroavi
I noticed that the author and reviewer guidelines state that the paper **may** exceed 1000 words:

> The paper should be between 250-1000 words. Authors submitting papers significantly longer than 1000 words may be asked to reduce the length of their paper.

So, if it is OK with you, we can address the issues above without cutting down on the current information.
