### 2023-09-17 TensorBFS/TensorInference.jl#66 @GiggleLiu
NOTE: The above link to tensor network definition has been updated.
https://tensorbfs.github.io/TensorInference.jl/dev/tensor-networks/
