### 2023-09-18 TensorBFS/TensorInference.jl#71 @gdalle
I don't have much experience as a JOSS reviewer, so I suggest you ask the editor directly in the review thread. Still, I think some content can easily be cut
