### 2023-09-22 TensorBFS/TensorInference.jl#63 @gdalle
this is non blocking for the review
