### 2023-09-22 TensorBFS/TensorInference.jl#67 @gdalle
much better!
