### 2023-09-22 TensorBFS/TensorInference.jl#66 @gdalle
I would just add one sentence to the paper defining a graphical model, before mentioning the treewidth
