### 2023-09-22 TensorBFS/TensorInference.jl#66 @gdalle
And mention that tensors are used to represent factors in a probability distribution
