### 2023-09-22 TensorBFS/TensorInference.jl#76 @gdalle
Code example in paper
There are a few details i don't understand in the code example:
- `In[4]`: when we assume the X-ray is _positive_, why does it correspond to `Dict(7 => 0)`
- `In[7]`: the phrasing is a bit ambiguous, there is only one most probable outcome which has several variables (and by the way, variables 3 and 4 are zero in this one)

https://github.com/openjournals/joss-reviews/issues/5700
