### 2023-09-22 #5700 @gdalle
As far as I'm concerned this is a significant improvement in clarity and accessibility, and we're nearly there!

- I suggest to merge the summary and statement of need in the paper, because at present they are better understood in reverse order (when reading first the statement of need and then the summary).
- I would like to see https://github.com/openjournals/joss-reviews/issues/5700 addressed or at least explained
