### 2023-09-23 b60141e
fix the overflow issue in probability
