+++ test/pr.jl
@@ -6,15 +6,15 @@ using TensorInference
 @testset "UAI Reference Solution Comparison" begin
     problems = dataset_from_artifact("uai2014")["PR"]
     problem_sets = [
-        #("Alchemy", TreeSA(ntrials = 1, niters = 5, βs = 0.1:0.1:100)), # fails
+        #("Alchemy", TreeSA(ntrials = 1, niters = 5, βs = 0.1:0.1:100)),
         #("CSP", TreeSA(ntrials = 1, niters = 5, βs = 0.1:0.1:100)),
         #("DBN", KaHyParBipartite(sc_target = 25)),
-        #("Grids", TreeSA(ntrials = 1, niters = 5, βs = 0.1:0.1:100)), # fails
-        #("linkage", TreeSA(ntrials = 3, niters = 20, βs = 0.1:0.1:40)), # fails
+        #("Grids", TreeSA(ntrials = 1, niters = 5, βs = 0.1:0.1:100)),
+        #("linkage", TreeSA(ntrials = 3, niters = 20, βs = 0.1:0.1:40)),
         #("ObjectDetection", TreeSA(ntrials = 1, niters = 5, βs = 1:0.1:100)),
         ("Pedigree", TreeSA(ntrials = 1, niters = 5, βs = 0.1:0.1:100)),
         #("Promedus", TreeSA(ntrials = 1, niters = 5, βs = 0.1:0.1:100)),
-        #("relational", TreeSA(ntrials=1, niters=5, βs=0.1:0.1:100)), # fails
+        #("relational", TreeSA(ntrials=1, niters=5, βs=0.1:0.1:100)), tw too large
         ("Segmentation", TreeSA(ntrials = 1, niters = 5, βs = 0.1:0.1:100))
     ]
     for (problem_set_name, optimizer) in problem_sets
