### 2023-09-24 457960f
Fix inconsistencies in the ASIA example
