### 2023-09-24 8457c6f
Fix inconsistencies in the example
Fixes: https://github.com/TensorBFS/TensorInference.jl/issues/76
