### 2023-09-24 TensorBFS/TensorInference.jl#76 @mroavi
Thanks for spotting these! You are right in both cases. We fixed them in the repo's example (https://github.com/TensorBFS/TensorInference.jl/commit/457960f9e1f7589e2f78f522d85e64dcbb11e493)
and in the paper (https://github.com/TensorBFS/TensorInference.jl/commit/8457c6f992a214d5c24f63060faa7bb03a57c786).
