### 2023-09-24 #5700 @mroavi
Hi @gdalle. Again, thanks for the suggestions to further improve the submission and the package itself.

> I suggest to merge the summary and statement of need in the paper, because at present they are better understood in reverse order (when reading first the statement of need and then the summary).

We did the exercise and found that it reads better in reverse order. We swapped the sections but kept them separate. This is because we want to keep these concerns separated and avoid having one very long section.

> I would like to see https://github.com/TensorBFS/TensorInference.jl/issues/65 addressed or at least explained

We finally had time to delve into this one. Please see our complete response under the [issue](https://github.com/TensorBFS/TensorInference.jl/issues/65) you opened. Here is a summary: The tests for both the MAR and PR tasks are now consistent with the reference solutions. We fixed an arithmetic overflow [here](https://github.com/TensorBFS/TensorInference.jl/pull/78), which was affecting our PR solutions. As for the MAP and MMAP tasks, we explain why we can't produce solutions for certain problems and why some of our results differ from those of other reference tools.
