### 2023-09-24 TensorBFS/TensorInference.jl#65 @gdalle
Sounds good to me! Thanks for the detailed rundown
