### 2023-09-24 #5700 @gdalle
I'm satisfied with this answer! And after checking the JOSS guidelines, the actual presence of a section named "statement of need"
is a requirement, so it is better not to merge
