### 2023-09-25 525a2b5
Create CUDA and GenericTensorNetworks extensions
