+++ src/TensorInference.jl
@@ -34,6 +34,10 @@ export sample
 # MMAP
 export MMAPModel
 
+# for GenericTensorNetworks
+export update_temperature
+function update_temperature end
+
 include("Core.jl")
 include("RescaledArray.jl")
 include("utils.jl")
@@ -41,12 +45,6 @@ include("mar.jl")
 include("map.jl")
 include("mmap.jl")
 include("sampling.jl")
-
-using Requires
-function __init__()
-    @require CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba" include("cuda.jl")
-    @require GenericTensorNetworks = "3521c873-ad32-4bb4-b63d-f4f178f42b49" include("generictensornetworks.jl")
-end
 
 # import PrecompileTools
 # PrecompileTools.@setup_workload begin
