### 2023-09-27 6b6c910
CompatHelper: bump compat for TropicalNumbers to 0.6, (keep existing compat) (#82)
Co-authored-by: CompatHelper Julia <compathelper_noreply@julialang.org>
