+++ Project.toml
@@ -24,5 +24,5 @@ OMEinsum = "0.7"
 PrecompileTools = "1"
 Requires = "1"
 StatsBase = "0.34"
-TropicalNumbers = "0.5.4"
+TropicalNumbers = "0.5.4, 0.6"
 julia = "1.3"
