### 2023-09-27 #5700 @mroavi
👋 @gdalle, I noticed that the check mark for the **automated tests** item is still missing from your list. I'm not sure if this is intentional or if something is still expected from our side.
