### 2023-10-02 TensorBFS/TensorInference.jl#84 @osorensen
JOSS submission: Reference check
Running our reference check suggests the following missing DOI. If the suggestion for the single missing DOI is correct, could you please update the BibTeX source file:

```

OK DOIs

- 10.1137/141000671 is OK
- 10.48550/arXiv.1209.5145 is OK
- 10.48550/arXiv.2108.05665 is OK
- 10.22331/q-2021-03-15-410 is OK
- 10.1103/PhysRevLett.128.030501 is OK
- 10.48550/ARXIV.2205.03718 is OK
- 10.1093/imaiai/iay009 is OK
- 10.1137/050644756 is OK
- 10.1016/j.aop.2014.06.013 is OK
- 10.1038/s42254-019-0086-7 is OK
- 10.1109/DSD57027.2022.00064 is OK
- 10.1145/567806.567807 is OK
- 10.1103/PhysRevB.99.155131 is OK
- 10.1103/PhysRevX.8.031012 is OK
- 10.48550/arXiv.2112.01657 is OK
- 10.1103/physrevlett.126.090506 is OK
- 10.1103/PhysRevX.9.031041 is OK
- 10.21468/SciPostPhysCodeb.4 is OK
- 10.5281/zenodo.8166121 is OK
- 10.7717/peerj-cs.1516 is OK
- 10.21105/joss.05161 is OK
- 10.1007/s10601-016-9245-y is OK

MISSING DOIs

- 10.1111/j.2517-6161.1988.tb01721.x may be a valid DOI for title: Local computations with probabilities on graphical structures and their application to expert systems

INVALID DOIs

- None
```

https://github.com/openjournals/joss-reviews/issues/5700
