### 2023-10-02 TensorBFS/TensorInference.jl#85 @osorensen
JOSS submission: Spelling in reference
On lines 70-71, do "Jin Guo Liu" and "Jin-Guo Liu" refer to the same person? If so, please spell consistently, so the references appears as "Liu et al., 2022; Liu et al. 2021".

https://github.com/openjournals/joss-reviews/issues/5700
