### 2023-10-02 TensorBFS/TensorInference.jl#86 @osorensen
JOSS submission: Small edits
On line 76, I suggest writing "KaHyParBipartite" and "SABipartite" without parentheses:

https://github.com/openjournals/joss-reviews/issues/5700
