### 2023-10-02 TensorBFS/TensorInference.jl#87 @osorensen
JOSS submission: consistent spelling
Please choose either "dyspnea" or "dyspnoea", and use it consistently. It seems like "dyspnea" is American English and "dyspnoea" is British English (https://en.wikipedia.org/wiki/Shortness_of_breath), so choose whichever you prefer.

https://github.com/openjournals/joss-reviews/issues/5700
