### 2023-10-02 TensorBFS/TensorInference.jl#88 @osorensen
Capitalize reference
Please make sure "Bayesian" is capitalized, by protecting the BibTeX entry with a set of extra curly braces:

https://github.com/openjournals/joss-reviews/issues/5700
