### 2023-10-02 e0a5d31
Add doi to lauritzen1988local bib entry
