+++ paper/paper.bib
@@ -117,7 +117,7 @@
 @misc{liu2022computing,
   doi = {10.48550/ARXIV.2205.03718},
   url = {https://arxiv.org/abs/2205.03718},
-  author = {Liu, Jin Guo and Gao, Xun and Cain, Madelyn and Lukin, Mikhail D. and Wang, Sheng-Tao},
+  author = {Liu, Jin-Guo and Gao, Xun and Cain, Madelyn and Lukin, Mikhail D. and Wang, Sheng-Tao},
   keywords = {Statistical Mechanics (cond-mat.stat-mech), FOS: Physical sciences, FOS: Physical sciences},
   title = {Computing solution space properties of combinatorial optimization problems via generic tensor networks},
   publisher = {arXiv},
@@ -186,7 +186,7 @@
 }
 
 @inproceedings{roa2023scaling,
-  author={Roa-Villescas, Martin and Liu, Jin Guo and Wijnings, Patrick W.A. and Stuijk, Sander and Corporaal, Henk},
+  author={Roa-Villescas, Martin and Liu, Jin-Guo and Wijnings, Patrick W.A. and Stuijk, Sander and Corporaal, Henk},
   booktitle={2023 Congress in Computer Science, Computer Engineering, & Applied Computing (CSCE)},
   title={Scaling Probabilistic Inference through Message Contraction Optimization},
   year={2023},
@@ -281,7 +281,7 @@
                   ho-oto and
                   maartenvd and
                   getzdan and
-                  Jinguo Liu and
+                  Jin-Guo Liu and
                   Dilum Aluthge and
                   Florian and
                   Spencer Lyon and
