+++ paper/figures/asia-network/asia-network.tex
@@ -164,7 +164,7 @@
     \texttt{B}                                              & Patient has bronchitis                   \\
     \texttt{E}                                              & Patient has \texttt{T} and/or \texttt{L} \\
     \texttt{X}                                              & Chest X-Ray is positive                  \\
-    \texttt{D}                                              & Patient has dyspnoea                     \\
+    \texttt{D}                                              & Patient has dyspnea                      \\
     \bottomrule()
   \end{longtable}
 \end{minipage}
