### 2023-10-02 817775d
Fix spelling errors and make casing in titles consistent
