+++ paper/paper.bib
@@ -1,5 +1,5 @@
 @article{bezanson2017julia,
-  title = {Julia: {A} {Fresh} {Approach} to {Numerical} {Computing}},
+  title = {Julia: A Fresh Approach to Numerical Computing},
   volume = {59},
   issn = {0036-1445},
   shorttitle = {Julia},
@@ -17,7 +17,7 @@
 }
 
 @article{bezanson2012julia,
-  title = {Julia: {A} {Fast} {Dynamic} {Language} for {Technical} {Computing}},
+  title = {Julia: A Fast Dynamic Language for Technical Computing},
   shorttitle = {Julia},
   url = {http://arxiv.org/abs/1209.5145},
   doi = {10.48550/arXiv.1209.5145},
@@ -148,18 +148,19 @@
   publisher={SIAM},
   doi={10.1137/050644756}
 }
+
 @article{orus2014practical,
-	doi = {10.1016/j.aop.2014.06.013},
-	year = 2014,
-	month = {oct},
-	publisher = {Elsevier {BV}
-},
-	volume = {349},
-	pages = {117--158},
-	author = {Rom{\'{a}}n Or{\'{u}}s},
-	title = {A practical introduction to tensor networks: Matrix product states and projected entangled pair states},
-	journal = {Annals of Physics},
-    doi={10.1016/j.aop.2014.06.013}
+  doi = {10.1016/j.aop.2014.06.013},
+  year = 2014,
+  month = {oct},
+  publisher = {Elsevier {BV}
+  },
+  volume = {349},
+  pages = {117--158},
+  author = {Rom{\'{a}}n Or{\'{u}}s},
+  title = {A practical introduction to tensor networks: Matrix product states and projected entangled pair states},
+  journal = {Annals of Physics},
+  doi={10.1016/j.aop.2014.06.013}
 }
 
 @article{orus2019tensor,
@@ -197,7 +198,7 @@
 }
 
 @article{blackford2002updated,
-  title={An updated set of basic linear algebra subprograms (BLAS)},
+  title={An updated set of basic linear algebra subprograms ({BLAS})},
   author={Blackford, L Susan and Petitet, Antoine and Pozo, Roldan and Remington, Karin and Whaley, R Clint and Demmel, James and Dongarra, Jack and Duff, Iain and Hammarling, Sven and Henry, Greg and others},
   journal={ACM Transactions on Mathematical Software},
   volume={28},
@@ -320,7 +321,7 @@
 }
 
 @article{oriol2023pymc,
-  title={PyMC: A Modern and Comprehensive Probabilistic Programming Framework in Python},
+  title={PyMC: A Modern and Comprehensive Probabilistic Programming Framework in {P}ython},
   author={Abril-Pla Oriol and Andreani Virgile and Carroll Colin and Dong Larry and Fonnesbeck Christopher J. and Kochurov Maxim and Kumar Ravin and Lao Jupeng and Luhmann Christian C. and Martin Osvaldo A. and Osthege Michael and Vieira Ricardo and Wiecki Thomas and Zinkov Robert},
   journal = {{PeerJ} Computer Science},
   publisher = {{PeerJ}},
@@ -353,7 +354,7 @@
   number = {84},
   pages = {5161},
   author = {Dmitry Bagaev and Albert Podusenko and Bert de Vries},
-  title = {RxInfer: A Julia package for reactive real-time Bayesian inference},
+  title = {RxInfer: A {J}ulia package for reactive real-time {B}ayesian inference},
   journal = {Journal of Open Source Software}
 }
 
@@ -377,17 +378,17 @@
 }
 
 @article{hurley2016multi,
[…]
