### 2023-10-02 367f16a
Improve description of contraction ordering methods
