### 2023-10-02 TensorBFS/TensorInference.jl#85 @mroavi
Nice catch! Should now be fixed by c982b5283b1e4272aa99b600ba5c1ecc9f2ba472.
