### 2023-10-02 TensorBFS/TensorInference.jl#86 @mroavi
Agreed. The clarity can be improved. Here is our attempt: 367f16a808468e5782cb3d35a53b266d04d77b82
