### 2023-10-02 TensorBFS/TensorInference.jl#87 @mroavi
We'll stick to American English. Here is the fix 8318f064b9b31087d670a06ade57631116cbad81
