### 2023-10-02 eaaea56
Fix invalid doi
