+++ paper/paper.bib
@@ -54,7 +54,7 @@
   number = {2},
   pages = {157-194},
   keywords = {artificial intelligence, bayesian methods, causal markov random field, decomposable graphs, expert systems, local potentials, markov random field, maximum cardinality search, probabilistic reasoning, triangulated graphs},
-  doi = {https://doi.org/10.1111/j.2517-6161.1988.tb01721.x},
+  doi = {10.1111/j.2517-6161.1988.tb01721.x},
   url = {https://rss.onlinelibrary.wiley.com/doi/abs/10.1111/j.2517-6161.1988.tb01721.x},
   eprint = {https://rss.onlinelibrary.wiley.com/doi/pdf/10.1111/j.2517-6161.1988.tb01721.x},
   abstract = {SUMMARY A causal network is used in a number of areas as a depiction of patterns of ‘influence’ among sets of variables. In expert systems it is common to perform ‘inference’ by means of local computations on such large but sparse networks. In general, non-probabilistic methods are used to handle uncertainty when propagating the effects of evidence, and it has appeared that exact probabilistic methods are not computationally feasible. Motivated by an application in electromyography, we counter this claim by exploiting a range of local representations for the joint probability distribution, combined with topological changes to the original network termed ‘marrying’ and ‘filling-in‘. The resulting structure allows efficient algorithms for transfer between representations, providing rapid absorption and propagation of evidence. The scheme is first illustrated on a small, fictitious but challenging example, and the underlying theory and computational aspects are then discussed.},
