### 2023-10-02 TensorBFS/TensorInference.jl#86 @osorensen
Very nice!
