### 2023-10-02 TensorBFS/TensorInference.jl#88 @mroavi
Fixed here 817775dd4f9afe02a5b28d93ffcd0a49087b0a01. We also improved the consistency of the casing of the titles.
