+++ v0.4.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"contributing/#Contributing","page":"Contributing","title":"Contributing","text":"","category":"section"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"Welcome and thank you for considering contributing to TensorInference.jl.","category":"page"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"By following these guidelines, you make it easier for everyone to work together. It shows you value the time of the people who create and manage this open-source project. In return, we'll show you the same respect by quickly looking at your issues, reviewing your changes, and helping you with your pull requests.","category":"page"},{"location":"contributing/#Getting-Started","page":"Contributing","title":"Getting Started","text":"","category":"section"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"Contributions are made to this repo via Issues and Pull Requests (PRs). A few general guidelines that cover both:","category":"page"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"Search for existing Issues and PRs before creating your own.\nWe do our best to solve problems quickly. Still, some issues take longer to understand. Posting a comment can help, especially if you need a quick fix.","category":"page"},{"location":"contributing/#Issues","page":"Contributing","title":"Issues","text":"","category":"section"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"Issues should be used to report problems with the package, ask for a new feature, or discuss possible changes before creating a Pull Request (PR). When you open a new issue, please include all the details needed to look into it.","category":"page"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"If you find an Issue that addresses the problem you're having, please add your own reproduction information to the existing issue rather than creating a new one. ","category":"page"},{"location":"contributing/#Pull-Requests","page":"Contributing","title":"Pull Requests","text":"","category":"section"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"Pull Requests (PRs) to our package are always welcome. Submitting a PR is a quick way to have your changes considered for the next software release. Generally, your Pull Request should:","category":"page"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"Either fix or add the functionality in question OR address widespread whitespace/style issues, not both.\nAdd unit or integration tests for any fixed or changed features.\nMinimize the number of changed lines to address a single concern.\nInclude documentation in the repo, especially for implementations of new features.","category":"page"},{"location":"contributing/","page":"Contributing","title":"Contributing","text":"For major changes that affect core functionality or would require a new major release, we recommend opening an Issue to discuss your ideas before creating a Pull Request. While this step is optional, it can sa
[…]
