### 2023-10-02 #5700 @mroavi
Version tag of the archived version: [v0.4.1](https://github.com/TensorBFS/TensorInference.jl/releases/tag/v0.4.1)
DOI of the archived version: 10.5281/zenodo.8399580
