### 2023-10-02 #5700 @osorensen
Thanks @mroavi. Could you please change the title of the Zenodo archive so it exactly matches that of the paper: "TensorInference: A Julia package for tensor-based probabilistic inference". This may result in a new DOI, if so, please post it here.
