### 2023-10-03 #5700 @mroavi
Paper accepted 🥳! A huge thanks to our editor @osorensen and to our reviewers @emstoudenmire and @gdalle. Your feedback helped improve our submission quite considerably. We appreciate the time and effort you've put into this. You've made this process not only constructive but also very enjoyable. We are grateful for your support for open-source projects and science in general.
