### 2023-10-12 #5700 @OriolAbril
Is it still possible to update the pdf?

It looks like there are some commas missing in https://github.com/TensorBFS/TensorInference.jl/blob/paper/paper/paper.bib#L325 and the reference is rendered using our full names and abbreviated surnames instead of the standard way
